+++ extensions/netbird/CHANGELOG.md
@@ -1,5 +1,10 @@
 # NetBird Changelog
 
+<<<<<<< HEAD
+## [Status on empty dns] - {PR_MERGE_DATE}
+
+- Fixed status command error when user does not have netbird DNS enabled
+
 ## [UX Improvements] - 2026-03-27
 
 - Status screen now shows empty screen during loading instead of empty data